requirements.txt
alabaster==0.7.12
alembic==1.0.5
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autoenv==1.0.0
autopep8==1.4.3
Babel==2.6.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cliff==2.14.0
cmd2
colorama==0.4.1
coverage==4.5.2
docutils==0.14
dominate==2.3.5
filelock==3.0.10
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
gnupg==2.3.1
gunicorn==19.9.0
idna==2.8
imagesize==1.1.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml
Mako
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe
more-itertools==5.0.0
numpy==1.15.4
packaging==18.0
pbr
Pillow
pluggy
prettytable
psutil
psycopg2-binary
py
pycodestyle
Pygments
pylint
pyparsing==2.3.0
pyperclip==1.7.0
pyreadline==2.1
pytest==4.0.2
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
python-secrets==18.11.5
pytz==2018.7
PyYAML==4.2b1 
requests==2.21.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.15
stevedore==1.30.0
toml==0.10.0
tox==3.6.0
typed-ast==1.1.1
urllib3==1.24.1
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
xkcdpass==1.16.5
yamlreader==3.0.4
